Adds first pass at medal review to decks screen
This commit adds a view of the different medals that you've earned to
the decks screen. These are not persisted between anki launches. Adds a
views.py file where the different components of the medals overview are
defined. Also adds the attrs library to make it easier to write decent
classes.