Adds vendor, adds database migrations
Adds a _vendor directory with a vendor.txt file that outlines the pypi
packages that should be installed and vendored. The vendor.py file
defines a routine that successfully installs vendored packages.

In terms of migrations, yoyo-migrations is used to handle the database.
The migrations are stored in the migrations directory, and they are run
when the main.py file is imported.