Moves state managment into ReviewingController
This commit makes the Store class, which holds the different "state
machines", return a new instance of the store class whenever the
on_answer etc. methods are called. The ReviewingController is passed an
instance of the store and holds and modifies the state. I like that it
is the only thing that is mutable right now.

The ReviewingController will also in the future take instances of the
AcheivementsRepository to add acheivements. Not sure yet what will
manage the medals view on the decks and other pages.