Gets initial persistence working
Repository now successfully loading all medals. Will want to use sql
group by for the medals count instead of what I'm doing now, and need to
change the fact that we are filtering to find the different displayable
medals, but otherwise this is working! Also should add an official medal
id field to the medals before launching.