Adds scoped medals overview to stats page
To add the scoped medals, this commit modifies the logic in main to
generate the correct header text and select the correct acheivements
from the acheivements repository. In the repository, this commit adds
methods to return the acheivements from deck ids and from the whole
collection. This probably could be refactored to not repeat the code as
much, but I'm not super interested in writing an ORM for this little
project.

This commit also corrects the pervasive typo where I spelled
achievements incorrectly.