Moves add-on toplevel state to profile_controller
This add-on refactors the main.py file to get ready for supporting
multiple profiles. To do this, I needed to make sure the method binding
and wrapping to the anki objects only occurs once. In doing this, we now
need to store global add-on state instead of just reinstanciating the
review controller everytime a new profile is loaded. This is required
because there is no easy way to unbind the method wrapping from the anki
objects.

The profile controller now holds the state that needs change whenever a
profile is loaded.