Sets up for mutliple profile handling
This commit rearranges and modifies a few things to get the add-on ready
to handle multiple profiles. First, it adds a decorator that makes sure the
profile controller is loaded before calling certain methods. This is
needed because anki calls profileLoaded hook AFTER the deck browser is
shown, so that without making sure the controller is loaded you wouldn't
see anything on the home screen the first time you open Anki.

Unfortunately, this necessitated getting rid of the thread calls for
injecting the js into the deck browser and stats screen. This shouldn't
be a big issue because those database calls are relatively optimized and
shouldn't take too long to return. It was nice while it lasted anyways.
Sqlite3/I is/am bad at handling connections and couldn't keep the
threading straight so I just got rid of it.

Next steps would be to pass the profile from anki into the load profile
method on the profile controller.