Adds killstreaks menu to bar, stores game_id in db
This commit is the first moves towards adding additional games to the
add-on. It creates a menu controller that holds a profile controller
that lets it talk to the db. Can probably rename profile controller to
profile provider and then have it just return profile specific state,
but that's a change for another time. Will rename menu controller in the
next commit to game controller.