Gets MW2 killstreaks in place
Adds a method to switch the game to the profile controller. The profile
controller now takes an object that maps game_id to the store for the
game. The store is what contains the state machines that track the
progress towards medals. Change game is called by the GameController,
which gest called whenever you switch the game at the top of the screen.