Updates medals overview to include all games
This commit updates the medals overview to include all games, and a
select box where you can choose what game you'll see the medals for. It
adds a few things to support that, including the jinja2 and markupsafe
libraries, a templates directory, and a js directory.

This commit moves the ad hoc templating from views.py into an actual
jinja2 template file, to make it easier to understand. It also adds a
medals_overview.js file that handles showing the correct medals for the
selected game.