Refactors profile cntrlr, removes reviewer delgtn
This commit adds a messy function to call a factory object, and then
call a method on the object created with given arguments. This lets us
not have to delegate the on_answer, on_show_question, on_show_answer
calls to the ReviewingController from the ProfileController. This turns
the ProfileController more into a provider of the current profile
dependent state (which is what it is intended to be, should probably
rename it)