Makes game auto switch!
This commit adds the ability to have the games automatically switch once
you've acheived all the medals for a game in a given session. To do
this, added a AllMedalsAcheivedNotifer class which wraps the
ReviewingController, subtracting each acheived medal from it's set of
medals until there are none left. Once that happens, it calls its passed
in notify function.