Fixes bug with game in menu and auto switching
Before, if you were using autoswitching and the game changed, the new
game selected in the Killstreaks menu. This commit fixes it by calling
game::set_current_game_id from within the profile controller, setting up
the callback to use ProfileController#change_game. Really need to test
the ProfileController, but the add-on might be done??