Makes all anki_killstreaks imports relative
This is necessary because when downloaded from ankiweb the top level
anki_killstreaks folder is unavailable. The only way to successfully
import things is through relative imports. This required many changes to
the _vendor files. Also, in the jinja compiler and environment files,
needed to make a couple changes to support everything being relative.