Refactors SettingsRepo db connection use
Commit refactors the SettingsRepo to use a factory function that gets a
db connection instead of taking a db connection. This is done because if
we want to send a SettingsRepo instance to another thread, it can't be
holding a db connection that was created in a different thread. This
will be useful whenever uploading achievement information to the
leaderboard. We will want that to happen in a different thread so the UI
does not hang.