Gets rid of css spinner and RetryingHttpClient
css spinner ended up taking up a ton of CPU and not being worth it. Just
prints a simple loading message now.

RetryingHttpClient did not work with the job queue as it is currently
written. Will need a different strategy to make sure that we upload when
people are offline.