Fixes bug where pressing numbers gives medals
This bug let you just spam the 2, 3, or 4 buttons on the question shown
screen and rack up achievements. Addon now checks for presence of right
hand reviews and whether or not that addon is enabled before allowing
you to get points just for hitting the numbers on the question shown
screen.